polishing and fixing
