+++ examples/bert/bert_pretrained_models/download_model.sh
@@ -1,2 +1,2 @@
-wget https://storage.googleapis.com/bert_models/2018_10_18/uncased_L-12_H-768_A-12.zip -P bert_released_models/;
-unzip bert_released_models/uncased_L-12_H-768_A-12.zip -d bert_released_models/
+wget https://storage.googleapis.com/bert_models/2018_10_18/uncased_L-12_H-768_A-12.zip -P bert_pretrained_models/;
+unzip bert_pretrained_models/uncased_L-12_H-768_A-12.zip -d bert_pretrained_models/
